+++ ffi/src/c_ffi/caci.rs
@@ -213,11 +213,11 @@ pub unsafe extern "C" fn tav_verify_caci_attestation(
         }?;
         let uvm_endorsement = unsafe { uvm_endorsement_handle(uvm_endorsement) }?;
         let uvm_feed = unsafe { input_text(uvm_feed, uvm_feed_len, "uvm_feed", false) }?;
-        let report_data = synchronous::verify_caci_attestation(
+        let report_data = synchronous::verify_caci_attestation_borrowed(
             *attestation,
             minimum_tcb,
             trusted_policy_digests,
-            uvm_endorsement.clone(),
+            uvm_endorsement,
             uvm_feed,
             minimum_svn,
         )